src/recordCreator/callRecord/requirements.txt
grpcio-tools ~= 1.30
opentelemetry-exporter-gcp-trace
opentelemetry-instrumentation-grpc